Enabled the entities package by changing to .net 4.0 and editing the manifest.json to allow enabling the package in package manager.
Removed old graphics from the project and replaced them. This includes flockers, their targets, and the environment. The old terrain was entirely removed and replaced by a plane.

Added a collision detection method for the separation behavior. This turns out more performant friendly if used when the separation force is high. Otherwise the previous method is more efficient.