Changed initializing settings on old managers.
Created scripts that implemented a job system version of both the regular and collision based flocking behaviors. Added new functions to the steering script for the job system that no longer utilize the character controller and also get the transform's forward and right vectors from the rotation. Also adds basic scripts used for testing basic movement with primitives using the job system to get a better understanding of how it works.

Added and edited prefabs and materials for various graphics elements including the grass, butterflies, the sky.