Adjustments:
Changed butterfly wing speed.
Changed materials for grass planes and skybox.
Changed camera to follow from the side instead of behind and adjusted height values.
Changed separation to take in a value for the distance to check to add a force.
Changed spawn positions ranges for the targets and flockers.
Changed texture sizes for the butterfly wings.
Made Camera a prefab in order to have multiple different scenes.
Removed unused textures.

New:

Imported 3d Game Kit in order to create an environment.
Added 3 more colors of butterflies. Added prefabs for each one.
Added different target prefabs for different showcase scenes.
Added 3 scenes for showcasing the project.
Added a script to display fps for the actual builds.
Added a script to use curves to move targets for the showcase.
Added a modified version of the manager that does not control the target.